requirements.txt
torch
datasets
tqdm
pyyaml
